fix: theme listener
