+++ src/palettes/changeTheme/index.js
@@ -2,11 +2,8 @@ import "./style.scss";
 import palette from "components/palette";
 import config from "lib/config";
 import appSettings from "lib/settings";
-import themes from "theme/list";
-import { updateSystemTheme } from "theme/preInstalled";
+import themes, { updateSystemThemeWatcher } from "theme/list";
 import changeEditorTheme from "../changeEditorTheme";
-
-const darkModeMediaQuery = window.matchMedia("(prefers-color-scheme: dark)");
 
 export default function changeTheme(type = "editor") {
 	if (type === "editor") return changeEditorTheme();
@@ -41,32 +38,6 @@ function generateHints(type) {
 		};
 	});
 }
-
-function syncSystemTheme(event) {
-	if (appSettings.value.appTheme.toLowerCase() !== "system") return;
-	const isDark = event ? event.matches : darkModeMediaQuery.matches;
-	updateSystemTheme(isDark);
-}
-
-function startSystemThemeWatcher() {
-	darkModeMediaQuery.addEventListener("change", syncSystemTheme);
-}
-
-function stopSystemThemeWatcher() {
-	darkModeMediaQuery.removeEventListener("change", syncSystemTheme);
-}
-
-function updateSystemThemeWatcher(theme) {
-	if (String(theme).toLowerCase() === "system") {
-		startSystemThemeWatcher();
-		syncSystemTheme();
-		return;
-	}
-	stopSystemThemeWatcher();
-}
-
-updateSystemThemeWatcher(appSettings.value.appTheme);
-appSettings.on("update:appTheme", updateSystemThemeWatcher);
 
 function onselect(value) {
 	if (!value) return;
